refactored FileNameMatcher into its own package
