+++ FileSearcher/src/FIleNameMatcher/BoundedPriorityQueue.java
@@ -1,4 +1,4 @@
-package SearchManager;
+package FIleNameMatcher;
 
 import java.util.Comparator;
 import java.util.Iterator;
