--- FileSearcher/src/SearchManager/FileTree.java
@@ -1,4 +1,0 @@
-package SearchManager;
-
-public class FileTree {
-}
